Jekyll hooks

+++ lib/jekyll/obsidian.rb
@@ -61,36 +61,49 @@ module Jekyll
         layouts_dir = File.join(File.dirname(site.dest), "_layouts")
         FileUtils.mkdir_p(layouts_dir) unless File.directory?(layouts_dir)
 
-        css_dir = File.join(File.dirname(site.dest), "assets", "obsidian", "css")
-        FileUtils.mkdir_p(css_dir) unless File.directory?(css_dir)
+        scss_dir = File.join(File.dirname(site.dest), "assets", "obsidian")
+        FileUtils.mkdir_p(scss_dir) unless File.directory?(scss_dir)
+
+        partials_dir = File.join(File.dirname(site.dest), "_sass", "obsidian")
+        FileUtils.mkdir_p(partials_dir) unless File.directory?(partials_dir)
 
         project_root = File.expand_path("../..", File.dirname(__FILE__))
-        assets_dir = File.join(project_root, "assets")
-        puts assets_dir
+        plugin_dir = File.join(project_root, "assets")
+        # puts plugin_dir
 
-        copy_files_from_dir(File.join(assets_dir, "css"), css_dir)
+        main_scss = File.join(plugin_dir, "css", "obsidian.scss")
+        copy_file_to_dir(main_scss, scss_dir)
 
-        copy_files_from_dir(File.join(assets_dir, "includes"), obsidian_dir)
+        copy_files_from_dir(File.join(plugin_dir, "css", "partials"), partials_dir)
 
-        layout = File.join(assets_dir, "layouts", "obsidian.html")
-        copy_file_to_dir(layout, layouts_dir)
+        layout = File.join(plugin_dir, "layouts", "obsidian.html")
+        copy_file_to_dir(layout, layouts_dir,true)
+
+        copy_files_from_dir(File.join(plugin_dir, "includes"), obsidian_dir,true)
       end
 
       private
 
-      def copy_file_to_dir(file, dir)
+      def copy_file_to_dir(file, dir, overwrite=false)
         if File.exist?(file)
-          FileUtils.cp(file, dir)
+          destination_file = File.join(dir, File.basename(file))
+
+          if !overwrite && File.exist?(destination_file)
+            puts "#{File.basename(file)} currently exists"
+          else
+            FileUtils.cp(file, dir)
+            puts "#{File.basename(file)} copied over"
+          end
         else
           puts "Error: #{file} does not exist"
           exit
         end
       end
 
-      def copy_files_from_dir(source_dir, destination_dir)
+      def copy_files_from_dir(source_dir, destination_dir, overwrite=false)
         Dir.glob(File.join(source_dir, '*')).each do |file_path|
-          next if File.directory?(file_path) # Skip directories
-          copy_file_to_dir(file_path, destination_dir)
+          next if File.directory?(file_path)
+          copy_file_to_dir(file_path, destination_dir,overwrite)
         end
       end
